Unify liveness behind one shared fork-free probe (internal/procinfo) (#87)
Codex review caught a second probe: internal/state had its OWN fork-based `ps`
matcher + EPERM-as-dead PIDAlive, feeding the no-session status BOARD and the
NOC console snapshots. So the cli fix alone left the board able to demote a live
agent under fork pressure — the surface amq-noc actually renders.

Extract ONE shared probe into internal/procinfo (a leaf package, no cycle) and
have BOTH internal/cli and internal/state use it:
- procinfo.Alive: signal-0 with EPERM => alive.
- procinfo.Match / procinfo.Args: fork-free command line (darwin KERN_PROCARGS2,
  linux /proc/<pid>/cmdline) with a ps -ww retry fallback.

cli's defaultDuplicateLaunchProbe and state's DefaultProbe/withDefaults now both
point at procinfo, so every surface (status/resume/doctor/preflight + board +
NOC snapshots) reads liveness identically. Probe tests moved to internal/procinfo.

Verified with ps sabotaged to always fail: single-session status, resume,
doctor, AND the no-session board all stay live (board 2/2 alive) — where the old
binary went stale.
