+++ .pm-os/state/proactive/pending-sessions.log
@@ -1,0 +1,9 @@
+2026-06-09T06:08:42%2B00:00|codex|019eaafe-0918-7b23-baf6-021d48623489|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T09-07-15-019eaafe-0918-7b23-baf6-021d48623489.jsonl||pending
+2026-06-09T06:15:14%2B00:00|codex|019eab03-c86b-7d20-8d51-391f57c06a48|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T09-13-32-019eab03-c86b-7d20-8d51-391f57c06a48.jsonl||pending
+2026-06-09T06:25:43%2B00:00|codex|019eab07-9d71-7162-a5cb-6056358f82d0|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T09-17-43-019eab07-9d71-7162-a5cb-6056358f82d0.jsonl||pending
+2026-06-09T06:34:23%2B00:00|codex|019eab14-46e1-7911-ae9b-9496f562f914|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T09-31-33-019eab14-46e1-7911-ae9b-9496f562f914.jsonl||pending
+2026-06-09T06:37:20%2B00:00|codex|019eab18-2491-74f3-a2e4-5bd5d231c047|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T09-35-46-019eab18-2491-74f3-a2e4-5bd5d231c047.jsonl||pending
+2026-06-09T06:55:22%2B00:00|codex|019eab25-6fd3-7190-bf77-151608f47146|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T09-50-17-019eab25-6fd3-7190-bf77-151608f47146.jsonl||pending
+2026-06-09T07:02:05%2B00:00|codex|019eab2c-ed1f-7ee2-8bef-e816ef4adc5b|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T09-58-28-019eab2c-ed1f-7ee2-8bef-e816ef4adc5b.jsonl||pending
+2026-06-09T08:14:44%2B00:00|codex|019eab6d-9590-75e2-b28f-72c16f62ec65|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T11-09-06-019eab6d-9590-75e2-b28f-72c16f62ec65.jsonl||pending
+2026-06-09T09:17:26%2B00:00|codex|019eaba9-0dc6-72c0-b847-dac9fb25f141|/Users/omri.a/Code/amq-squad|/Users/omri.a/.codex/sessions/2026/06/09/rollout-2026-06-09T12-14-03-019eaba9-0dc6-72c0-b847-dac9fb25f141.jsonl||pending
