--- internal/cli/procargs_linux.go
@@ -1,18 +1,0 @@
-//go:build linux
-
-package cli
-
-import (
-	"os"
-	"strconv"
-)
-
-// readProcArgsNative reads pid's full command line from /proc/<pid>/cmdline —
-// no fork, so it does not fail under fork/exec pressure the way `ps` does (#87).
-func readProcArgsNative(pid int) (string, bool) {
-	buf, err := os.ReadFile("/proc/" + strconv.Itoa(pid) + "/cmdline")
-	if err != nil {
-		return "", false
-	}
-	return parseProcCmdline(buf)
-}
